Move alert rule name to the top left in alert form, and add a placeholder
