+++ public/app/features/alerting/unified/components/rule-editor/DetailsStep.tsx
@@ -19,12 +19,6 @@ import { RuleEditorSection } from './RuleEditorSection';
 import { RuleFolderPicker, Folder, containsSlashes } from './RuleFolderPicker';
 import { checkForPathSeparator } from './util';
 
-const recordingRuleNameValidationPattern = {
-  message:
-    'Recording rule name must be valid metric name. It may only contain letters, numbers, and colons. It may not contain whitespace.',
-  value: /^[a-zA-Z_:][a-zA-Z0-9_:]*$/,
-};
-
 interface DetailsStepProps {
   initialFolder: RuleForm | null;
 }
@@ -46,7 +40,7 @@ export const DetailsStep = ({ initialFolder }: DetailsStepProps) => {
 
   return (
     <RuleEditorSection
-      stepNo={type === RuleFormType.cloudRecording ? 2 : 3}
+      stepNo={type === RuleFormType.cloudRecording ? 3 : 4}
       title={
         type === RuleFormType.cloudRecording ? 'Add details for your recording rule' : 'Add details for your alert'
       }
@@ -56,31 +50,6 @@ export const DetailsStep = ({ initialFolder }: DetailsStepProps) => {
           : 'Write a summary and add labels to help you better manage your alerts'
       }
     >
-      <Field
-        className={styles.formInput}
-        label="Rule name"
-        error={errors?.name?.message}
-        invalid={!!errors.name?.message}
-      >
-        <Input
-          id="name"
-          {...register('name', {
-            required: { value: true, message: 'Must enter an alert name' },
-            pattern: ruleFormType === RuleFormType.cloudRecording ? recordingRuleNameValidationPattern : undefined,
-            validate: {
-              pathSeparator: (value: string) => {
-                // we use the alert rule name as the "groupname" for Grafana managed alerts, so we can't allow path separators
-                if (ruleFormType === RuleFormType.grafana) {
-                  return checkForPathSeparator(value);
-                }
-
-                return true;
-              },
-            },
-          })}
-        />
-      </Field>
-
       {(ruleFormType === RuleFormType.cloudRecording || ruleFormType === RuleFormType.cloudAlerting) &&
         dataSourceName && <GroupAndNamespaceFields rulesSourceName={dataSourceName} />}
