+++ public/app/features/alerting/unified/components/rule-editor/GrafanaEvaluationBehavior.tsx
@@ -91,7 +91,7 @@ export const GrafanaEvaluationBehavior = () => {
 
   return (
     // TODO remove "and alert condition" for recording rules
-    <RuleEditorSection stepNo={2} title="Alert evaluation behavior">
+    <RuleEditorSection stepNo={3} title="Alert evaluation behavior">
       <Field
         label="Evaluate"
         description="Evaluation interval applies to every rule within a group. It can overwrite the interval of an existing alert rule."
