+++ public/app/features/alerting/unified/components/rule-editor/NotificationsStep.tsx
@@ -5,8 +5,8 @@ import { useFormContext } from 'react-hook-form';
 import { GrafanaTheme2 } from '@grafana/data';
 import { Card, Link, useStyles2, useTheme2 } from '@grafana/ui';
 
-import { RuleFormValues } from '../../types/rule-form';
 import { GRAFANA_RULES_SOURCE_NAME } from '../../utils/datasource';
+import { RuleFormType, RuleFormValues } from '../../types/rule-form';
 
 import LabelsField from './LabelsField';
 import { RuleEditorSection } from './RuleEditorSection';
@@ -15,14 +15,15 @@ export const NotificationsStep = () => {
   const [hideFlowChart, setHideFlowChart] = useState(false);
   const styles = useStyles2(getStyles);
   const theme = useTheme2();
+  const { watch } = useFormContext<RuleFormValues & { location?: string }>();
 
-  const { watch } = useFormContext<RuleFormValues>();
+  const type = watch('type');
 
   const dataSourceName = watch('dataSourceName') ?? GRAFANA_RULES_SOURCE_NAME;
 
   return (
     <RuleEditorSection
-      stepNo={4}
+      stepNo={type === RuleFormType.cloudRecording ? 4 : 5}
       title="Notifications"
       description="Grafana handles the notifications for alerts by assigning labels to alerts. These labels connect alerts to contact points and silence alert instances that have matching labels."
     >
